add work with MongoDB + fixes
