--- src/main/java/nflx/bench/CreditPipeline/dto/TaskResponse.java
@@ -1,13 +1,0 @@
-package nflx.bench.CreditPipeline.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class TaskResponse {
-    private int code;
-    private String message;
-}
