--- src/main/java/nflx/bench/CreditPipeline/repository/TaskRepository.java
@@ -1,15 +1,0 @@
-package nflx.bench.CreditPipeline.repository;
-
-import nflx.bench.CreditPipeline.dto.Task;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface TaskRepository extends MongoRepository<Task, UUID> {
-
-    Optional<Task> findById(UUID id);
-    List<Task> findAll();
-
-}
