+++ src/main/java/nflx/bench/creditpipeline/dto/Task.java
@@ -1,18 +1,19 @@
-package nflx.bench.CreditPipeline.dto;
+package nflx.bench.creditpipeline.dto;
 
 import lombok.*;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.util.UUID;
 
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
+
+@Data
+@Document
 public class Task {
     @Id
     private UUID id;
+
     private String title;
+
     private Boolean isCompleted;
 }
